requirements.txt
aerich~=0.6.3
aioredis~=2.0.0
asyncpg==0.28.0
cryptography~=43.0.0
fastapi[standard]~=0.98.0
starlette~=0.27.0
pydantic~=1.10.11
pyjwt~=2.9.0
tortoise-orm~=0.19.0
uvicorn[standard]~=0.22.0